feat(types): move Asset type to asset-service
BREAKING CHANGE: Asset is now exported from asset-service, and all consumers should now import it from it
