+++ packages/asset-service/src/generateAssetData/ethereum/uniswapV2Pools.ts
@@ -1,6 +1,6 @@
 import { ethChainId as chainId, toAssetId } from '@shapeshiftoss/caip'
-import { Asset } from '@shapeshiftoss/types'
 
+import { Asset } from '../../service/AssetService'
 import { ethereum } from '../baseAssets'
 import { colorMap } from '../colorMap'
