+++ packages/asset-service/src/generateAssetData/ethereum/yearnVaults.ts
@@ -1,10 +1,10 @@
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { ethChainId as chainId, toAssetId } from '@shapeshiftoss/caip'
-import { Asset } from '@shapeshiftoss/types'
 import { Token, Vault } from '@yfi/sdk'
 import { Yearn } from '@yfi/sdk'
 import toLower from 'lodash/toLower'
 
+import { Asset } from '../../service/AssetService'
 import { ethereum } from '../baseAssets'
 import { colorMap } from '../colorMap'
