+++ packages/asset-service/src/generateAssetData/generateAssetData.ts
@@ -1,11 +1,10 @@
 import 'dotenv/config'
 
 import { CHAIN_REFERENCE, fromAssetId } from '@shapeshiftoss/caip'
-import { Asset } from '@shapeshiftoss/types'
 import fs from 'fs'
 import orderBy from 'lodash/orderBy'
 
-import { AssetsById } from '../service/AssetService'
+import { Asset, AssetsById } from '../service/AssetService'
 import * as avalanche from './avalanche'
 import { atom, bitcoin } from './baseAssets'
 import * as ethereum from './ethereum'
