+++ packages/asset-service/src/generateAssetData/utils.ts
@@ -1,6 +1,6 @@
-import { Asset } from '@shapeshiftoss/types'
 import filter from 'lodash/filter'
 
+import { Asset } from '../service/AssetService'
 import blacklist from './blacklist.json'
 
 export const filterOutBlacklistedAssets = (unfilteredAssetData: Asset[]) =>
