+++ packages/swapper/src/api.ts
@@ -1,14 +1,9 @@
+import { Asset } from '@shapeshiftoss/asset-service'
 import { AssetId, ChainId } from '@shapeshiftoss/caip'
 import { avalanche, cosmos, ethereum, osmosis } from '@shapeshiftoss/chain-adapters'
 import { createErrorClass } from '@shapeshiftoss/errors'
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import {
-  Asset,
-  BIP44Params,
-  ChainSpecific,
-  KnownChainIds,
-  UtxoAccountType
-} from '@shapeshiftoss/types'
+import { BIP44Params, ChainSpecific, KnownChainIds, UtxoAccountType } from '@shapeshiftoss/types'
 
 export const SwapError = createErrorClass('SwapError')
